Move "refresh quota" logic to repository
