+++ app/src/main/kotlin/at/bitfire/davdroid/ui/webdav/WebdavMountsModel.kt
@@ -5,37 +5,30 @@
 package at.bitfire.davdroid.ui.webdav
 
 import android.app.Application
-import android.provider.DocumentsContract
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import at.bitfire.davdroid.R
 import at.bitfire.davdroid.db.AppDatabase
 import at.bitfire.davdroid.db.WebDavMount
 import at.bitfire.davdroid.webdav.CredentialsStore
 import at.bitfire.davdroid.webdav.DavDocumentsProvider
+import at.bitfire.davdroid.webdav.WebDavMountRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 @HiltViewModel
 class WebdavMountsModel @Inject constructor(
     val context: Application,
-    val db: AppDatabase
+    val db: AppDatabase,
+    val mountRepository: WebDavMountRepository
 ): ViewModel() {
 
-    val authority = context.getString(R.string.webdav_authority)
-
-    val mountDao = db.webDavMountDao()
-    val documentDao = db.webDavDocumentDao()
-
-    val mounts = mountDao.getAllFlow()
-    val mountInfos = mountDao.getAllWithRootDocumentFlow()
+    private val mounts = mountRepository.getAllFlow()
+    val mountInfos = mountRepository.getAllWithRootFlow()
 
     var refreshingQuota by mutableStateOf(false)
         private set
@@ -58,24 +51,7 @@ class WebdavMountsModel @Inject constructor(
         refreshingQuota = true
 
         viewModelScope.launch {
-            withContext(Dispatchers.IO) {
-                val resolver = context.contentResolver
-                mountDao.getAll().forEach { mount ->
-                    documentDao.getOrCreateRoot(mount).let { root ->
-                        var loading = true
-                        while (loading) {
-                            val rootDocumentUri = DocumentsContract.buildChildDocumentsUri(authority, root.id.toString())
-                            resolver.query(rootDocumentUri, null, null, null, null)?.use { cursor ->
-                                loading = cursor.extras.getBoolean(DocumentsContract.EXTRA_LOADING)
-                            }
-
-                            if (loading)        // still loading, wait a bit
-                                delay(100)
-                        }
-                    }
-                }
-            }
-
+            mountRepository.refreshAllQuota()
             refreshingQuota = false
         }
     }
